Fixes en las reseñas y aspecto renovado. Refs #12

+++ frontend/app/src/main/java/com/example/playtracker/data/repository/ReviewsRepository.kt
@@ -6,7 +6,7 @@ import com.example.playtracker.domain.model.Review
 interface ReviewsRepository {
     suspend fun upsert(
         gameId: Long,
-        score0to10: Int?,
+        score0to10: Float?,
         notes: String?,
         containsSpoilers: Boolean = false,
         bearer: String
